Make cpufrequtils optional
cpufrequtils is not available on some architectures,
such as PowerPC machines.
In these cases, previously the charm would fail to install,
crashing on trying to install cpufrequtils.

So now the charm will install cpufrequtils if it is available
and continue as usual.
If cpufrequtils is not available however,
the charm will skip attempting to install cpufrequtils,
and ignore the governor config option.