Make cpufrequtils optional (#90)
cpufrequtils is not available on some architectures,
such as ppc64el.
In these cases, previously the charm would fail to install,
crashing on trying to install cpufrequtils.

So now the charm will install cpufrequtils if it is available
and continue as usual.
If cpufrequtils is not available however,
the charm will skip attempting to install cpufrequtils,
and ignore the governor config option.

Also some tox.ini improvements:
- remove the envdir option from tox.ini;
  this is not a recommended way to share environments.
- switch to flake8-pyproject
  - appears more maintained
  - lets us use the original flake8 command